adapter/server: moved request into server adapter
